requirements.txt
polars==0.20.11
python-dotenv==1.0.0
pandas==2.2.1
SQLAlchemy==2.0.27
pyodbc==5.1.0
pyarrow==15.0.0